refactor(table): move table atoms to its folder
